Remove `constexpr inline` across proj, and other changes
A constexpr function is already inline

+++ src/ADT/FileDescriptor.cpp
@@ -15,7 +15,7 @@
 
 FileDescriptor
 FileDescriptor::Open(const char *Path, OpenKind Kind, Flags Flags) noexcept {
-    return open(Path, static_cast<int>(Kind), Flags);
+    return open(Path, static_cast<int>(Kind), static_cast<int>(Flags));
 }
 
 FileDescriptor FileDescriptor::Create(const char *Path, int Mode) noexcept {
